[CodeGen][Passes] Move `CodeGenPassBuilder.h` to Passes
